+++ tket/include/Utils/MatrixAnalysis.hpp
@@ -175,6 +175,6 @@ bool in_weyl_chamber(const std::array<Expr, 3> &k);
 /**
  * @brief Get an nth root of a 2x2 unitary matrix.
  */
-Eigen::Matrix2cd nth_root(const Eigen::Matrix2cd &u, unsigned n);
+Eigen::Matrix2cd nth_root(const Eigen::Matrix2cd &u, unsigned long long n);
 
 }  // namespace tket
