+++ tket/src/Utils/MatrixAnalysis.cpp
@@ -528,7 +528,7 @@ bool in_weyl_chamber(const std::array<Expr, 3> &k) {
   return true;
 }
 
-Eigen::Matrix2cd nth_root(const Eigen::Matrix2cd &u, unsigned n) {
+Eigen::Matrix2cd nth_root(const Eigen::Matrix2cd &u, unsigned long long n) {
   if (u.isApprox(Eigen::Matrix2cd::Identity(), EPS)) {
     return Eigen::Matrix2cd::Identity();
   } else if (n == 0) {
